Display new logo in ANSI image with Pixterm
