+++ plank/pkg/server/server.go
@@ -447,66 +447,6 @@ func (ps *platformServer) configureRobotsPath() {
 			})))
 }
 
-// printBanner prints out banner as well as brief server config
-func (ps *platformServer) printBanner() {
-	fmt.Println(` ______ __      ______  __   __  __  __    
-/\  == /\ \    /\  __ \/\ "-.\ \/\ \/ /    
-\ \  _-\ \ \___\ \  __ \ \ \-.  \ \  _"-.  
- \ \_\  \ \_____\ \_\ \_\ \_\\"\_\ \_\ \_\ 
-  \/_/   \/_____/\/_/\/_/\/_/ \/_/\/_/\/_/ 
-                                           `)
-	utils.Infof("Host\t\t\t")
-	fmt.Println(ps.serverConfig.Host)
-	utils.Infof("Port\t\t\t")
-	fmt.Println(ps.serverConfig.Port)
-
-	if ps.serverConfig.FabricConfig != nil {
-		utils.Infof("Fabric endpoint\t\t")
-		fmt.Println(ps.serverConfig.FabricConfig.FabricEndpoint)
-	}
-
-	if len(ps.serverConfig.StaticDir) > 0 {
-		utils.Infof("Static endpoints\t")
-		for i, dir := range ps.serverConfig.StaticDir {
-			_, p := utils.DeriveStaticURIFromPath(dir)
-			fmt.Print(p)
-			if i < len(ps.serverConfig.StaticDir)-1 {
-				fmt.Print(", ")
-			} else {
-				fmt.Print("\n")
-			}
-		}
-	}
-
-	if ps.serverConfig.SpaConfig != nil {
-		utils.Infof("SPA endpoint\t\t")
-		fmt.Println(ps.serverConfig.SpaConfig.BaseUri)
-		utils.Infof("SPA static assets\t")
-		if len(ps.serverConfig.SpaConfig.StaticAssets) == 0 {
-			fmt.Print("-")
-		}
-		for idx, asset := range ps.serverConfig.SpaConfig.StaticAssets {
-			_, uri := utils.DeriveStaticURIFromPath(asset)
-			fmt.Print(utils.SanitizeUrl(uri, false))
-			if idx < len(ps.serverConfig.SpaConfig.StaticAssets)-1 {
-				fmt.Print(", ")
-			}
-		}
-		fmt.Println()
-	}
-
-	utils.Infof("Health endpoint\t\t")
-	fmt.Println("/health")
-
-	if ps.serverConfig.EnablePrometheus {
-		utils.Infof("Prometheus endpoint\t")
-		fmt.Println("/prometheus")
-	}
-
-	fmt.Println()
-
-}
-
 func (ps *platformServer) checkPortAvailability() {
 	// is the port free?
 	_, err := net.Dial("tcp", fmt.Sprintf(":%d", ps.serverConfig.Port))
